✨CubeMXでプロジェクト生成して、FLASH書き込み読み出しができたのと、AS5048Aを読み取れた
- stm32pioでPlatformIOのプロジェクトに変更している
- HAL_SPI_TransmitReceive_DMAで0xFFFFに投げて、その次に空データ送りつけて読み込みをDMAでやっている
- DMA、めっちゃ複雑なんかと思ったけど、CubeMXでの設定はめっちゃ簡単だった。DMAと普通のやつの違いがよくわからん。DMAを使うとリングバッファ的なやつを作れるのかな。今回そんなことしてないけど。
- DMAで16bit送るようにしているっぽい。できているのだろう。
- AS5048Aを12bitと勘違いしていたが14bitでした。2^14(0x3FFF)でマスクしようね。
- mbedだとFLASHの書き込み読み出しが上手いことできなかった(mbedが立ち上がるとFFに書き換えられる問題があった)が、HALでフルスクラッチになったらFFに書き換えが起こらなかったのでラッキー

+++ BLDC_programV2/lib/BLDC/BLDC.cpp
@@ -169,11 +169,11 @@ float BLDCMotor::updateEncoder() {
     }
 }
 
-int BLDCMotor::getShaftAngle() {
+float BLDCMotor::getShaftAngle() {
     return shAngle;
 }
 
-int BLDCMotor::getElectricAngle() {
+float BLDCMotor::getElectricAngle() {
     return elAngle;
 }
